end list, insert

+++ src/main/java/com/tm/cgv/movieImage/MovieImageRepository.java
@@ -10,7 +10,7 @@ public interface MovieImageRepository {
 	public List<MovieImageVO> movieImageList(int movieNum)throws Exception;
 	
 	//이미지사진이 여기까지 들어와야함.
-	public int movieImageInsert(MovieImageVO movieImageVO) throws Exception;
+	public long  movieImageInsert(MovieImageVO movieImageVO) throws Exception;
 	
 	public MovieImageVO movieImageSelect(MovieImageVO movieImageVO) throws Exception;
